Force epoch count to int

+++ models/autoencoders.py
@@ -97,7 +97,7 @@ class Generic_AE(nn.Module):
         config = {}
         config['optim']     = optim.Adam(self.parameters(), lr=1e-3)
         config['scheduler'] = optim.lr_scheduler.ReduceLROnPlateau(config['optim'], patience=10, threshold=1e-3, min_lr=1e-6, factor=0.1, verbose=True)
-        config['max_epoch'] = 240 * self.epoch_factor
+        config['max_epoch'] = int(240 * self.epoch_factor)
         return config
 
     def preferred_name(self):
